[rename] directory organization
