--- vite-project/src/App.jsx
@@ -1,12 +1,0 @@
-import "./App.css";
-import AppRoutes from "./Router.jsx";
-
-function App() {
-  return (
-    <div className="App">
-      <AppRoutes />
-    </div>
-  );
-}
-
-export default App;
