+++ vite-project/src/CSS/Footer.css
@@ -1,5 +1,3 @@
-@import "./styleguide.css";
-
 .footer {
   height: 70px;
   padding: 45px 100px;
