+++ vite-project/src/Pages/Footer.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./Footer.css";
+import "../CSS/Footer.css";
 
 const Footer = () => {
   return (
