+++ vite-project/src/Pages/main.jsx
@@ -1,14 +1,10 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
+import '../index.css'
 import App from './App.jsx'
-import Header from './Header.jsx'
-import Footer from './Footer.jsx'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <Header />
     <App />
-    <Footer />
   </StrictMode>,
 )
